Use Compose/Docker clients directly rather than Remote API calls
Previously we were using the Python Docker client library, dockerpty-py,
and the compose library in the test rig because these give us a little
bit more "direct" API access. But in addition to these being not very
good libraries, it fails us from a testing standpoint. We don't *want*
to test the Remote API (which is presumably well-tested by our core
engineering team), we want to test using the actual client binaries that
Docker, Inc. provides.

So instead of using these libs to make API calls as we were, this commit
makes `subprocess.Popen` calls to Docker or Compose. A nice added
advantage is that we'll be able to test against multiple versions of the
client in parallel if we want.