Allow `docker` and `docker-compose` binary to be overridden
Uses the `COMPOSE` and `DOCKER` env vars to override the path to the
binary for these two external applications. Refactored the many
separate `subprocess.check_output` calls to Docker into a single method
so that we can make sure we use the override and ensure these calls
have been instrumented.